+++ src/abstract/DbQueryPool.ts
@@ -1,8 +1,9 @@
 import type { DbCursor, DbCursorOption } from "./DbCursor.ts";
 import { DbQuery } from "./DbQuery.ts";
-import type { DbPool, DbPoolConnection, DbTransaction, SqlLike, TransactionMode } from "./interfaces.ts";
+import type { DbPool, DbPoolConnection, DbTransaction, SqlLike } from "./interfaces.ts";
 import type { QueryRowsResult } from "./DbQueryBase.ts";
-import type { InferQueryResult, SqlStatementDataset } from "./external.ts";
+import type { SqlStatementDataset } from "./external.ts";
+import type { InferQueryResult, TransactionMode } from "@/query.ts";
 
 /**
  * @public