+++ src/mod.ts
@@ -2,7 +2,6 @@ export * from "./impl/mod.ts";
 export * from "./util.ts";
 export * from "./abstract/mod.ts";
 
-export * from "./pg_oid.ts";
 export * from "./connect.ts";
 export * from "./query.ts";
 export * from "./platforms.ts";