+++ src/query.ts
@@ -2,3 +2,5 @@ export * from "./query/QueryStatement.ts";
 export * from "./query/MessageData.ts";
 export * from "./query/data_decoder.ts";
 export * from "./query/QueryReader.ts";
+export * from "@/query/Query.ts";
+export * from "./query/Transaction.ts";