+++ test/tests/abstract/type.test.ts
@@ -1,4 +1,5 @@
-import type { InferQueryResult, SqlStatementDataset } from "#abstract";
+import type { SqlStatementDataset } from "#abstract";
+import type { InferQueryResult } from "@/query.ts";
 import { expectTypeOf, test } from "vitest";
 
 test("推断类型", function () {